Move Clock element from logic to timers namespace
